pyproject.toml
[tool.poetry]
name = "etl-pipeline"
version = "0.1.0"
description = "Schema-driven self-extending ETL pipeline with two-agent architecture"
authors = ["aq"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2"
langchain = "^0.3"
langchain-core = "^0.3"
langchain-anthropic = "^0.3"
langgraph = "^0.4"
litellm = "^1.55"
python-dotenv = "^1.0"
rapidfuzz = "^3.9"
sentence-transformers = "^3.0"
faiss-cpu = "^1.8"
numpy = "^1.26"
streamlit = "^1.56.0"
tqdm = "^4.0"
pyyaml = "^6.0"
structlog = "^24.0"
boto3 = "^1.42.91"
gcloud = "^0.18.3"
google-cloud-storage = "^3.10.1"
google-cloud-bigquery = "^3.41.0"
db-dtypes = "^1.5.1"
redis = "^7.4.0"
chromadb = ">=0.6"
pyarrow = "^16.0"
rank-bm25 = "^0.2"
mlxtend = "^0.23"
networkx = "^3.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.pytest.ini_options]
testpaths = ["tests", "src"]
markers = [
    "integration: requires real GCS credentials (deselect with '-m not integration')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
